requirements.txt
Django==1.3
django-annoying==0.7.6
django-uni-form==0.8.0
